Task is now linked with user
Voter system implement for acl
Auto-assign with logged user
Wrote and refactor Tests (still have to work on)

Resolves #1, resolves #10, resolves #11 and resolves #12

+++ src/AppBundle/Controller/TaskController.php
@@ -92,6 +92,7 @@ class TaskController extends Controller
      */
     public function deleteTaskAction(Task $task)
     {
+        $this->denyAccessUnlessGranted('delete', $task);
         $em = $this->getDoctrine()->getManager();
         $em->remove($task);
         $em->flush();
